requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Flask-WTF==1.2.2
Pillow==11.1.0
python-dotenv==1.0.1
Werkzeug==3.1.3
pandas==2.2.3
openpyxl==3.1.5
gunicorn==23.0.0
psycopg2-binary==2.9.10
